[feat] Studio 프로필 이미지 validation 추가

+++ yp-online/src/main/java/kr/co/yourplanet/online/business/studio/service/impl/ProfileServiceImpl.java
@@ -70,6 +70,12 @@ public class ProfileServiceImpl implements ProfileService {
         if (studioDto.isDuplicateIds()) {
             throw new BusinessException(StatusCode.BAD_REQUEST, "중복된 포트폴리오 ID가 포함되어 있습니다.", false);
         }
+
+        // 프로필 이미지 변경 되었을 경우
+        if (Boolean.TRUE.equals(studioDto.getProfileImageChanged())) {
+            fileManageUtil.validateImageFile(profileImageFile);
+        }
+
         Member member = memberRepository.findById(memberId).orElseThrow(() -> new BusinessException(StatusCode.NOT_FOUND, "해당 회원 정보가 존재하지 않습니다.", false));
         Optional<Studio> optionalStudio = studioRepository.findById(memberId);
         Studio studio = optionalStudio.orElseGet(() -> Studio.builder()
@@ -113,14 +119,15 @@ public class ProfileServiceImpl implements ProfileService {
             portfolioCategoryMapRepository.deleteAllByStudioAndCreateDateBeforeAndUpdateDateBefore(studio, now, now);
         }
 
-        // 모든 스튜디오 프로필 저장 로직이 완료된 후 프로필 이미지 저장처리
-        // 기존에 존재하는 프로필 이미지 삭제
-        if (StringUtils.hasText(studio.getProfileImagePath())) {
-            fileManageUtil.deleteFile(studio.getProfileImagePath());
-            studio.updateProfileImage("", "");
-        }
-        // 프로필 이미지 파일시스템 저장
-        if (profileImageFile != null && !profileImageFile.isEmpty()) {
+        // 프로필 이미지 저장 부분
+        if (Boolean.TRUE.equals(studioDto.getProfileImageChanged())) {
+            // 모든 스튜디오 프로필 저장 로직이 완료된 후 프로필 이미지 저장처리
+            // 기존에 존재하는 프로필 이미지 삭제
+            if (StringUtils.hasText(studio.getProfileImagePath())) {
+                fileManageUtil.deleteFile(studio.getProfileImagePath());
+                studio.updateProfileImage("", "");
+            }
+            // 프로필 이미지 파일시스템 저장
             FileUploadResult uploadResult = fileManageUtil.uploadFile(profileImageFile, FileType.PROFILE_IMAGE);
             studio.updateProfileImage(uploadResult.getFilePath(), uploadResult.getFileUrl());
         }
